requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
python-multipart==0.0.18
jinja2==3.1.5
google-api-python-client==2.149.0
google-auth-oauthlib==1.2.1
Pillow==11.1.0
